fix: Extract SystemFragment (#239)
* Extract SystemFragment

* Move system child fragments to shared
