+++ src/service/impl/fragments/mobility-gql/shared.graphql
@@ -32,3 +32,43 @@ fragment translation on Translation {
     language
     value
 }
+
+fragment rentalUris on RentalUris {
+    android
+    ios
+}
+
+fragment rentalApp on RentalApp {
+    discoveryUri
+    storeUri
+}
+
+fragment rentalApps on RentalApps {
+    android {
+        ...rentalApp
+    }
+    ios {
+        ...rentalApp
+    }
+}
+
+fragment brandAssets on BrandAssets {
+    brandImageUrl
+    brandImageUrlDark
+    brandLastModified
+}
+
+fragment system on System {
+    operator {
+        ...operator
+    }
+    name {
+        ...translatedString
+    }
+    brandAssets {
+        ...brandAssets
+    }
+    rentalApps {
+        ...rentalApps
+    }
+}
