+++ src/service/impl/fragments/mobility-gql/stations.graphql-gen.ts
@@ -1,12 +1,10 @@
 import * as Types from '../../../../graphql/mobility/mobility-types_v2';
 
-import { TranslatedStringFragment, PricingPlanFragment, OperatorFragment } from './shared.graphql-gen';
-import { BrandAssetsFragment, RentalAppsFragment, RentalUrisFragment } from './vehicles.graphql-gen';
+import { TranslatedStringFragment, PricingPlanFragment, SystemFragment, RentalUrisFragment } from './shared.graphql-gen';
 import { DocumentNode } from 'graphql';
 import gql from 'graphql-tag';
-import { TranslatedStringFragmentDoc, PricingPlanFragmentDoc, OperatorFragmentDoc } from './shared.graphql-gen';
-import { BrandAssetsFragmentDoc, RentalAppsFragmentDoc, RentalUrisFragmentDoc } from './vehicles.graphql-gen';
-export type StationFragment = { id: string, lat: number, lon: number, capacity?: number, numBikesAvailable: number, numDocksAvailable?: number, name: TranslatedStringFragment, pricingPlans: Array<PricingPlanFragment>, system: { operator: OperatorFragment, name: TranslatedStringFragment, brandAssets?: BrandAssetsFragment, rentalApps?: RentalAppsFragment }, rentalUris?: RentalUrisFragment };
+import { TranslatedStringFragmentDoc, PricingPlanFragmentDoc, SystemFragmentDoc, RentalUrisFragmentDoc } from './shared.graphql-gen';
+export type StationFragment = { id: string, lat: number, lon: number, capacity?: number, numBikesAvailable: number, numDocksAvailable?: number, name: TranslatedStringFragment, pricingPlans: Array<PricingPlanFragment>, system: SystemFragment, rentalUris?: RentalUrisFragment };
 
 export const StationFragmentDoc = gql`
     fragment station on Station {
@@ -23,18 +21,7 @@ export const StationFragmentDoc = gql`
     ...pricingPlan
   }
   system {
-    operator {
-      ...operator
-    }
-    name {
-      ...translatedString
-    }
-    brandAssets {
-      ...brandAssets
-    }
-    rentalApps {
-      ...rentalApps
-    }
+    ...system
   }
   rentalUris {
     ...rentalUris
@@ -42,9 +29,7 @@ export const StationFragmentDoc = gql`
 }
     ${TranslatedStringFragmentDoc}
 ${PricingPlanFragmentDoc}
-${OperatorFragmentDoc}
-${BrandAssetsFragmentDoc}
-${RentalAppsFragmentDoc}
+${SystemFragmentDoc}
 ${RentalUrisFragmentDoc}`;
 export type Requester<C = {}, E = unknown> = <R, V>(doc: DocumentNode, vars?: V, options?: C) => Promise<R>
 export function getSdk<C, E>(requester: Requester<C, E>) {
