+++ src/service/impl/fragments/mobility-gql/vehicles.graphql-gen.ts
@@ -1,20 +1,12 @@
 import * as Types from '../../../../graphql/mobility/mobility-types_v2';
 
-import { TranslatedStringFragment, PricingPlanFragment, OperatorFragment } from './shared.graphql-gen';
+import { TranslatedStringFragment, PricingPlanFragment, SystemFragment, RentalUrisFragment } from './shared.graphql-gen';
 import { DocumentNode } from 'graphql';
 import gql from 'graphql-tag';
-import { TranslatedStringFragmentDoc, PricingPlanFragmentDoc, OperatorFragmentDoc } from './shared.graphql-gen';
+import { TranslatedStringFragmentDoc, PricingPlanFragmentDoc, SystemFragmentDoc, RentalUrisFragmentDoc } from './shared.graphql-gen';
 export type VehicleTypeFragment = { id: string, formFactor: Types.FormFactor, maxRangeMeters?: number, name?: TranslatedStringFragment };
 
-export type RentalUrisFragment = { android?: string, ios?: string };
-
-export type RentalAppFragment = { discoveryUri?: string, storeUri?: string };
-
-export type RentalAppsFragment = { android?: RentalAppFragment, ios?: RentalAppFragment };
-
-export type BrandAssetsFragment = { brandImageUrl: string, brandImageUrlDark?: string, brandLastModified: string };
-
-export type VehicleFragment = { id: string, lat: number, lon: number, isReserved: boolean, isDisabled: boolean, currentRangeMeters: number, currentFuelPercent?: number, availableUntil?: string, vehicleType: VehicleTypeFragment, pricingPlan: PricingPlanFragment, system: { operator: OperatorFragment, name: TranslatedStringFragment, brandAssets?: BrandAssetsFragment, rentalApps?: RentalAppsFragment }, rentalUris?: RentalUrisFragment };
+export type VehicleFragment = { id: string, lat: number, lon: number, isReserved: boolean, isDisabled: boolean, currentRangeMeters: number, currentFuelPercent?: number, availableUntil?: string, vehicleType: VehicleTypeFragment, pricingPlan: PricingPlanFragment, system: SystemFragment, rentalUris?: RentalUrisFragment };
 
 export const VehicleTypeFragmentDoc = gql`
     fragment vehicleType on VehicleType {
@@ -26,35 +18,6 @@ export const VehicleTypeFragmentDoc = gql`
   }
 }
     ${TranslatedStringFragmentDoc}`;
-export const BrandAssetsFragmentDoc = gql`
-    fragment brandAssets on BrandAssets {
-  brandImageUrl
-  brandImageUrlDark
-  brandLastModified
-}
-    `;
-export const RentalAppFragmentDoc = gql`
-    fragment rentalApp on RentalApp {
-  discoveryUri
-  storeUri
-}
-    `;
-export const RentalAppsFragmentDoc = gql`
-    fragment rentalApps on RentalApps {
-  android {
-    ...rentalApp
-  }
-  ios {
-    ...rentalApp
-  }
-}
-    ${RentalAppFragmentDoc}`;
-export const RentalUrisFragmentDoc = gql`
-    fragment rentalUris on RentalUris {
-  android
-  ios
-}
-    `;
 export const VehicleFragmentDoc = gql`
     fragment vehicle on Vehicle {
   id
@@ -72,18 +35,7 @@ export const VehicleFragmentDoc = gql`
     ...pricingPlan
   }
   system {
-    operator {
-      ...operator
-    }
-    name {
-      ...translatedString
-    }
-    brandAssets {
-      ...brandAssets
-    }
-    rentalApps {
-      ...rentalApps
-    }
+    ...system
   }
   rentalUris {
     ...rentalUris
@@ -91,10 +43,7 @@ export const VehicleFragmentDoc = gql`
 }
     ${VehicleTypeFragmentDoc}
 ${PricingPlanFragmentDoc}
-${OperatorFragmentDoc}
-${TranslatedStringFragmentDoc}
-${BrandAssetsFragmentDoc}
-${RentalAppsFragmentDoc}
+${SystemFragmentDoc}
 ${RentalUrisFragmentDoc}`;
 export type Requester<C = {}, E = unknown> = <R, V>(doc: DocumentNode, vars?: V, options?: C) => Promise<R>
 export function getSdk<C, E>(requester: Requester<C, E>) {
